Rename file utils.rb to jira-input-plugin-utils.rb
