+++ lib/embulk/input/jira-input-plugin-utils.rb
